fix: project folder structure
