+++ index.ts
@@ -1,0 +1,5 @@
+import routerDex from "./lib/routerDex"
+
+console.log("module")
+
+export default routerDex
